S3 Bam Restoration Feature (#229)
